Increased stack size for larger jpeg images. Abstracted jpeg decoding into own module
