+++ Jpeg_RGB.h
@@ -1,0 +1,11 @@
+#ifndef __JPEG_RGB_H
+#define __JPEG_RGB_H
+
+typedef struct RGB
+{
+  uint8_t B;
+  uint8_t G;
+  uint8_t R;
+}RGB_typedef;
+
+#endif /* __JPEG_RGB_H */
