+++ RTE/Device/STM32F746NGHx/startup_stm32f746xx.s
@@ -45,9 +45,7 @@
 ;   <o> Stack Size (in Bytes) <0x0-0xFFFFFFFF:8>
 ; </h>
 
-;Stack_Size      EQU     0x00000400
-;Stack_Size      EQU     0x2000
-Stack_Size      EQU     0xF000
+Stack_Size      EQU     0x0000F000
 
                 AREA    STACK, NOINIT, READWRITE, ALIGN=3
 Stack_Mem       SPACE   Stack_Size
@@ -58,9 +56,7 @@ __initial_sp
 ;   <o>  Heap Size (in Bytes) <0x0-0xFFFFFFFF:8>
 ; </h>
 
-;Heap_Size       EQU     0x00000200
-Heap_Size        EQU	 0xF000
-;Heap_Size       EQU		0x2F000
+Heap_Size       EQU     0x0000F000
 
                 AREA    HEAP, NOINIT, READWRITE, ALIGN=3
 __heap_base
