+++ SDCard_Module.c
@@ -14,6 +14,7 @@
 #include "Camera_Defines.h"
 #include "Board_GLCD.h"
 #include "GLCD_Config.h"
+#include "Jpeg_Decode.h"
 
 extern GLCD_FONT     GLCD_Font_16x24;
 
@@ -187,14 +188,13 @@ int SDCard_GetBMPFileName(const char* DirName, char* Files[], const unsigned int
   return index;
 }
 
-#include "Camera_Globals.h"
-extern unsigned char *buffer;
-uint8_t   _aucLine[256];
-uint32_t  offset = 0;
-RGB_typedef *RGB_matrix;
-static uint8_t processBuffer(uint8_t* Row, uint32_t rowNum, uint16_t width, uint16_t height);
+//#include "Camera_Globals.h"
+//extern unsigned char *buffer;
+//uint32_t  offset = 0;
+//RGB_typedef *RGB_matrix;
+//static uint8_t processBuffer(uint8_t* Row, uint32_t rowNum, uint16_t width, uint16_t height);
 
-int SDCard_loadJPEG(unsigned char *buff, const char* filePath, uint16_t *imgWidth,uint16_t *imgHeight)
+/*int SDCard_loadJPEG(unsigned char *buff, const char* filePath, uint16_t *imgWidth, uint16_t *imgHeight)
 {
 	FIL file;
 
@@ -212,34 +212,37 @@ int SDCard_loadJPEG(unsigned char *buff, const char* filePath, uint16_t *imgWidt
 	}
 
 	//
-	jpeg_decode(&file, imgWidth, imgHeight, _aucLine, processBuffer);
+	jpeg_decode(&file, imgWidth, imgHeight);
+	//jpeg_decode(&file, imgWidth, imgHeight, _aucLine, processBuffer);
 	
 	f_close (&file);
 	f_mount(NULL, (TCHAR const*)"",0);
 	
 	return 1;
+}*/
+
+int SDCard_OpenFile(FIL *file, const char *filePath)
+{
+	  // Open filesystem
+  if(f_mount(&fs, (TCHAR const*)"",0) != FR_OK)
+  {
+    return 0;
+  }
+	
+	// Open file
+	FRESULT res = f_open(file, (TCHAR const*)filePath, FA_READ);
+	if( res != FR_OK)
+  {
+		return 0;
+	}
+	
+	return 1;
 }
 
-static uint8_t processBuffer(uint8_t* Row, uint32_t rowNum, uint16_t width, uint16_t height)
+int SDCard_CloseFile(FIL *file)
 {
-	offset = ((uint32_t)buffer + (width * (height - rowNum) * 2));
-	RGB_matrix =  (RGB_typedef*)Row;
-	int index = 0;
-	for (index = 0; index < width; index++)
-  {
-		// Convert RGB888 to RGB565.
-		// Also swap RB.
-		uint8_t Red16   = RGB_matrix[index].B  >> 3;  // 5-bit red
-    uint8_t Green16 = RGB_matrix[index].G  >> 2;  // 6-bit green
-    uint8_t Blue16  = RGB_matrix[index].R  >> 3;  // 5-bit blue
-		
-		unsigned short col = (unsigned short) Blue16 + (Green16<<5) + (Red16<<(5+6));
-
-		*(__IO uint8_t*) (offset) = *((uint8_t *)&col);
-		offset++;
-		*(__IO uint8_t*) (offset) = *((uint8_t *)&col+1); 
-		offset++;
-	}  
+	f_close (file);
+	f_mount(NULL, (TCHAR const*)"",0);
 	
 	return 1;
 }
